fix: restructure overflowItems in separate folder
