+++ packages/ibm-products/src/global/js/hooks/useOverflowItems/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export { useOverflowItems } from './useOverflowItems';
